+++ src/server/service/creacteNoteFromActivityService/index.ts
@@ -1,0 +1,1 @@
+export * as createNoteActivityService from "./service";
